INC-126 refactor: extract common logic from Followers and Following to hook
