+++ src/components/pagesComponents/admin/adminUserPage/followers/Followers.tsx
@@ -1,27 +1,11 @@
-import { useEffect, useState } from 'react'
-
 import { useCommonTablePagination } from '@/common/hooks/useCommonTablePagination'
-import { Column, CommonTable } from '@/components/ui/commonTable'
+import { useFollow } from '@/common/hooks/useFollow'
+import { CommonTable } from '@/components/ui/commonTable'
 import { Pagination } from '@/components/ui/pagination'
-import { Follow } from '@/services/admin/types'
-import { useGetFollowersQuery, useGetUserLazyQuery } from '@/services/admin/usersService.generated'
+import { useGetFollowersQuery } from '@/services/admin/usersService.generated'
 import { useRouter } from 'next/router'
 
-type FollowersWithFullNames = ({ fullName: string } & Follow)[]
-
 export const Followers = () => {
-  const columns: Column<FollowersWithFullNames[number]>[] = [
-    { accessor: 'userId', title: 'User ID' },
-    { accessor: 'fullName', title: 'Username' },
-    {
-      accessor: 'userName',
-      href: row => `/profile/${row.userId}`,
-      isLink: true,
-      sortable: true,
-      title: 'Profile link',
-    },
-    { accessor: 'createdAt', sortable: true, title: 'Subscription Date' },
-  ]
   const router = useRouter()
   const { query } = router
   const userId = Number(query.id)
@@ -32,32 +16,10 @@ export const Followers = () => {
     variables: { pageNumber, pageSize, userId: 1 },
   })
 
-  const [getFollowerFullName] = useGetUserLazyQuery()
-  const [followersWithFullNames, setFollowersWithFullNames] = useState<FollowersWithFullNames>([])
-  const followers = followersData?.getFollowers.items
+  const followers = followersData?.getFollowers.items ?? []
   const totalCount = followersData?.getFollowers.totalCount
 
-  useEffect(() => {
-    const fetchFollowersFullNames = async () => {
-      if (!followers) {
-        return
-      }
-      const followersWithFullNames = await Promise.all(
-        followers.map(async follower => {
-          const { data } = await getFollowerFullName({ variables: { userId: follower.id } })
-          const firstName = data?.getUser.profile.firstName ?? ''
-          const lastName = data?.getUser.profile.lastName ?? ''
-          const fullName = `${firstName} ${lastName}`
-
-          return { ...follower, fullName }
-        })
-      )
-
-      setFollowersWithFullNames(followersWithFullNames)
-    }
-
-    fetchFollowersFullNames()
-  }, [followers, getFollowerFullName])
+  const { columns, itemsWithFullNames: followersWithFullNames } = useFollow({ items: followers })
 
   return (
     <>
