+++ src/components/pagesComponents/admin/adminUserPage/following/Following.tsx
@@ -1,27 +1,11 @@
-import { useEffect, useState } from 'react'
-
 import { useCommonTablePagination } from '@/common/hooks/useCommonTablePagination'
-import { Column, CommonTable } from '@/components/ui/commonTable'
+import { useFollow } from '@/common/hooks/useFollow'
+import { CommonTable } from '@/components/ui/commonTable'
 import { Pagination } from '@/components/ui/pagination'
-import { Follow } from '@/services/admin/types'
-import { useGetFollowingQuery, useGetUserLazyQuery } from '@/services/admin/usersService.generated'
+import { useGetFollowingQuery } from '@/services/admin/usersService.generated'
 import { useRouter } from 'next/router'
 
-type FollowingWithFullNames = ({ fullName: string } & Follow)[]
-
 export const Following = () => {
-  const columns: Column<FollowingWithFullNames[number]>[] = [
-    { accessor: 'userId', title: 'User ID' },
-    { accessor: 'fullName', title: 'Username' },
-    {
-      accessor: 'userName',
-      href: row => `/profile/${row.userId}`,
-      isLink: true,
-      sortable: true,
-      title: 'Profile link',
-    },
-    { accessor: 'createdAt', sortable: true, title: 'Subscription Date' },
-  ]
   const router = useRouter()
   const { query } = router
   const userId = Number(query.id)
@@ -33,32 +17,10 @@ export const Following = () => {
     variables: { pageNumber, pageSize, userId: 7 },
   })
 
-  const [getFollowingFullName] = useGetUserLazyQuery()
-  const [followingWithFullNames, setFollowingWithFullNames] = useState<FollowingWithFullNames>([])
-  const following = followingData?.getFollowing.items
+  const following = followingData?.getFollowing.items ?? []
   const totalCount = followingData?.getFollowing.totalCount
 
-  useEffect(() => {
-    const fetchFollowingsFullNames = async () => {
-      if (!following) {
-        return
-      }
-      const followingWithFullNames = await Promise.all(
-        following.map(async followingItem => {
-          const { data } = await getFollowingFullName({ variables: { userId: followingItem.id } })
-          const firstName = data?.getUser.profile.firstName ?? ''
-          const lastName = data?.getUser.profile.lastName ?? ''
-          const fullName = `${firstName} ${lastName}`
-
-          return { ...followingItem, fullName }
-        })
-      )
-
-      setFollowingWithFullNames(followingWithFullNames)
-    }
-
-    fetchFollowingsFullNames()
-  }, [following, getFollowingFullName])
+  const { columns, itemsWithFullNames: followingWithFullNames } = useFollow({ items: following })
 
   return (
     <>
